Spark: move actions to a separate package
